Speed improvements to TSV savings shape calculations
When a measure has a custom 8760 savings shape CSV attached to it, all savings shapes for each combination of region, building type, and end use for the measure are retrieved once when the measure is initialized, rather than being retrieved each time a measure microsegment is processed.

Additionally, EnergyPlus prototype names were adjusted slightly in the TSV load data and cost conversion data.