Fixed sector-level 8760 overwrite issue
When multiple EMM regions were specified in an ECM’s definition with the 8760 sector-level shapes option (--sect_shapes), successive regions in the microsegment updating loop would write over previous regions’ sector shapes data; this issue was fixed.

Additionally, a bug in a screening for valid region names in the measure JSON was fixed.