Handle incorrect region setting, add options to docs
When a user specifies TSV ECM features and/or evaluation metrics but has not specified the analysis regions as EMM, ecm_prep.py now does so automatically while warning the user about the change.

ECM preparation options are all documented in a subsection of Local Execution Tutorial 2 in the docs.